Add custom tls
Modifies the client.mustache template to include a tls fix the error
`local error: tls: unexpected message` that is received when making any
calls with the generated base Go library. This was an issue in our
Atrium library as well and we fixed it there. This includes
the same fix in our Platform API Go library.

Atrium fix:
https://github.com/mxenabled/atrium-go/pull/40